color.rb: fix a bug that display wrong color for FalseClass in 1.8, see:
since hash should be unordered, and even if it's order preserved,
what if a user inserts their own color? at any rate, this color
looking up should be ordered by subclass < superclass, that is,
we should always check if the most subclass match or not, and if not,
check next. Object should be checked last.

since ruby provide Class#<, i think it should provide <=> too... :(