squeeze_history.rb: fix multiline + squeeze_history bug, see:
when we're doing multiline editing, we might input:
"
"
which would be squeezed but it's a valid input.
what here we should do is, squeeze_history should process last,
after multiline_history had fixed the history, so that it should be
\n"\n"
and this won't be squeezed. squeeze_history should override
loop_once, instead of anything related to input