only require core for bin/rib, not for Rib.shell, see:
at first i want to give people core plugins automatically
if you're not using any plugin. so that using `rib` directly
would give you an irb like session. but this is buggy :(

if there's a program:

    require 'rib/more/anchor'
    Rib.anchor 123

then this would crash, because we ended up with plugins:
(crash reason is that Anchor is not overriding Underscore!)

    [Rib::Anchor, Rib::Underscore]

but it should be:

    [Rib::Underscore, Rib::Anchor]

or merely:

    [Rib::Anchor]

we can't really detect if we would want core plugins when
calling `Rib.shell`, so this is too messy!

thanks andrew