make requiring plugins after the shell launched work, see detail:
because if you are requiring `require 'rib/extra/autoindent'`
after entering the shell, the `before_loop` for the Autoindent
plugin won't be called, thus ended up with the configuration is nil.

now we access the configuration through a method, so that we can
lazily assign defaults while `before_loop` didn't get called at all.