fix a bug where you run the shell in top level, while anchor...
in the other source, and exit the anchored shell, it would simply
stop from anchored place and jump back to the top level. this is
because my naive implementation to avoid printing from leaving
the anchored shell. this shouldn't be done in a throw/catch
manner, but a special return value to check whether it is returned
from anchored.