actually, using TOPLEVEL_BINDING would make backtrace more clear
however, too bad we can't create <main> method as before!
if we can, then we could make it as if we were actually on
TOPLEVEL_BINDING, not a jailed one.

then how did irb solve this? no, it only strips irb_binding
from backtrace... and i don't want to do that, either. so
here we just try to name it main, and remove it afterward
to avoid polluting TOPLEVEL_BINDING

p.s. local variables are not shared for this, either