Breaking changes: use nested modules
Instead of using `module Rib::Color`, use
`module Rib; module Color` because of constant
resolution. Before so, we'll need to `include Rib`
in order to see constants in Rib, but that's somehow a
bad enforcement. It would be nice that we don't have to
include Rib to see constants in Rib.