Fence and restore ctrl+c interruption
Consider you're anchoring in rspec, or maybe unicorn,
normally you hit ctrl+c to clear the input, but under
rspec and unicorn, they would trap SIGINT and overwrite
the default behaviour, causing it to quit early while
you're still using rib.

This is really annoying. I need to remember don't hit
ctrl+c in those cases.

Now we trap again each time a shell starts running,
overwriting rspec or unicorn, so that we could clear
the input normally. After we're done with the shell,
we of course need to restore the original SIGINT handler,
so that could be back to work.