[more/color] Fix when the input is next or break
When input is next or break, backtrace would be simply `(rib)` without a
colon (:), therefore msgs would be nil, causing errors. Not sure why,
but when that happens we could just append the line number there.

The test (test_shell.rb) can't catch this directly, but when it's
running with the other plugins like `env TEST_LEVEL=2`, it could
catch it.